requirements.txt
pydicom==2.4.0
pillow==10.2.0